updated previous code from this pr #1019
